+++ src/pages/MergePositions/Contents.tsx
@@ -1,25 +1,25 @@
+import { getLogger } from 'util/logger'
+import { isConditionFullIndexSet, minBigNumber } from 'util/tools'
+import { Status } from 'util/types'
+
+import { Button } from 'components/buttons/Button'
+import { CenteredCard } from 'components/common/CenteredCard'
+import { Amount } from 'components/mergePositions/Amount'
+import { MergePreview } from 'components/mergePositions/MergePreview'
+import { SelectCondition } from 'components/mergePositions/SelectCondition'
+import { SelectPosition } from 'components/mergePositions/SelectPosition'
+import { ButtonContainer } from 'components/pureStyledComponents/ButtonContainer'
+import { Row } from 'components/pureStyledComponents/Row'
+import { FullLoading } from 'components/statusInfo/FullLoading'
+import { IconTypes } from 'components/statusInfo/common'
+import { ZERO_BN } from 'config/constants'
+import { useConditionContext } from 'contexts/ConditionContext'
+import { useMultiPositionsContext } from 'contexts/MultiPositionsContext'
+import { useWeb3Connected } from 'contexts/Web3Context'
 import { ethers } from 'ethers'
 import { BigNumber } from 'ethers/utils'
 import React, { useCallback, useMemo, useState } from 'react'
-
-import { Button } from '../../components/buttons/Button'
-import { CenteredCard } from '../../components/common/CenteredCard'
-import { Amount } from '../../components/mergePositions/Amount'
-import { MergePreview } from '../../components/mergePositions/MergePreview'
-import { SelectCondition } from '../../components/mergePositions/SelectCondition'
-import { SelectPosition } from '../../components/mergePositions/SelectPosition'
-import { ButtonContainer } from '../../components/pureStyledComponents/ButtonContainer'
-import { Row } from '../../components/pureStyledComponents/Row'
-import { FullLoading } from '../../components/statusInfo/FullLoading'
-import { IconTypes } from '../../components/statusInfo/common'
-import { ZERO_BN } from '../../config/constants'
-import { useConditionContext } from '../../contexts/ConditionContext'
-import { useMultiPositionsContext } from '../../contexts/MultiPositionsContext'
-import { useWeb3Connected } from '../../contexts/Web3Context'
-import { ConditionalTokensService } from '../../services/conditionalTokens'
-import { getLogger } from '../../util/logger'
-import { isConditionFullIndexSet, minBigNumber } from '../../util/tools'
-import { Status } from '../../util/types'
+import { ConditionalTokensService } from 'services/conditionalTokens'
 
 const logger = getLogger('MergePosition')
